Fixed Hunter in village generation. All hunters are currently spawning in the villages center, I tried to change that as well, but it probably would take a lot of time to do and test that.
Also the generation does not work for custom villages yet.

+++ src/main/java/de/teamlapen/vampirism/util/MobProperties.java
@@ -11,5 +11,5 @@ public class MobProperties {
 	public static final double vampireHunter_maxHealth = 20.0;
 	public static final double vampireHunter_attackDamage = 5.0;
 	public static final double vampireHunter_movementSpeed = 0.4;
-	public static final double vampireHunter_maxPerVillage = 5;
+	public static final int vampireHunter_maxPerVillage = 5;
 }
